fix(#2229): stop truncating 64-bit connect and idle deltas to int
CLinearTimeDelta::ReturnSeconds() is 64-bit, but the reporting path narrowed it
to int, so a delta past 2^31 wrapped into a plausible smaller number instead of
an obviously wrong one. That is how #2227 stayed hidden: a live descriptor whose
conn(<port>) is 6,220,867,076 s -- 197 years -- shows in WHO as 22290d 11h,
which is exactly that value less 2^32. Sixty-one years reads as an oddity; a
hundred and ninety-seven would have been reported long ago.

Widened time_format_1, time_format_2, expand_time, write_time and
GeneralTimeConversion, along with fetch_connect and fetch_idle and their
callers. GeneralTimeConversion's int iValue accumulator had to go with them --
with a 64-bit input the term-by-term remainder subtraction overflowed and every
term printed the whole span rather than what was left of it. Its buffer-size
comments already reason in terms of 2^63, so 64-bit was the original intent.

fun_singletime, fun_exptime and fun_digittime narrowed at the softcode entry
point too, so singletime(6220596985) answered 61y.

Two details worth calling out:

  - tf1_width_table used INT_MAX as the end-of-cases sentinel for widths 10 and
    11.  Widening the parameter alone would let a large value walk past the
    final case into case 3, whose format spec is empty at those widths.  The
    real range maxima from the comment block replace it.

  - A value too large for the column count now yields "???", which the function
    already emits for a bad maxWidth.  Rendering an impossible duration as a
    believable one is the failure being fixed; a marker is the point.

Verified against a live server by injecting the #2227 corruption:

  singletime(6220596985)   61y            -> 197y
  exptime(6220596985)      cumulative junk -> 197y 1M 2w 12h 20s
  WHO On For, width 10     22290d 11h     -> 71997d 15h
  WHO On For, width 11     (wrapped)      -> 155397d 15h
  WHO On For, INT64_MIN/2  (garbage)      -> ???

197y 1M 2w 12h 20s decomposes back to 6,220,596,985 exactly.  Ordinary values
are unchanged: singletime(90061) is 1d and a fresh session still reads 0:00.

+++ mux/src/netcommon.cpp
@@ -1357,7 +1357,7 @@ int fetch_width(dbref target)
 // ---------------------------------------------------------------------------
 // fetch_idle: Return smallest idle time for a player (or -1 if not logged in).
 //
-int fetch_idle(dbref target)
+int64_t fetch_idle(dbref target)
 {
     CLinearTimeAbsolute ltaNow;
     ltaNow.GetUTC();
@@ -1403,7 +1403,7 @@ void find_oldest(const dbref target, DESC *dOldest[2])
 // fetch_connect: Return largest connect time for a player (or -1 if not
 // logged in).
 //
-int fetch_connect(dbref target)
+int64_t fetch_connect(dbref target)
 {
     DESC *dOldest[2];
     find_oldest(target, dOldest);
